Adjust failing test & refactoring.
